+++ app/poll/model.py
@@ -1,7 +1,8 @@
 from typing import Optional, List
 from sqlmodel import SQLModel, Field, Relationship
-from ..store.model import Store_Info
-from ..user.model import User
+from app.store.model import Store_Info
+from app.user.model import User
+from app.store.model import Store_Info
 
 
 # 투표 모델
@@ -40,15 +41,3 @@ class Vote(SQLModel):
     user_id: List["User"] = Relationship(back_populates='vote')
 
 
-from sqlmodel import SQLModel, create_engine
-
-sqlite_file_name = "./test_database.db"
-sqlite_url = f"sqlite:///{sqlite_file_name}"
-
-engine = create_engine(sqlite_url, echo=True)
-
-def create_db_and_tables():
-    SQLModel.metadata.create_all(engine)
-
-if __name__ == "__main__":
-    create_db_and_tables()
