+++ app/store/menu/model.py
@@ -1,3 +1,4 @@
+
 from typing import Optional
 from sqlmodel import Field, SQLModel, Relationship
